[bug] Patch the windows menu bar regressions
