+++ src/main/menu.about.ts
@@ -3,7 +3,7 @@ import { BrowserWindow, clipboard, dialog, MenuItemConstructorOptions } from "el
 import { I18nService } from "jslib-common/abstractions/i18n.service";
 
 import { UpdaterMain } from "jslib-electron/updater.main";
-import { isMac, isSnapStore, isWindowsStore } from "jslib-electron/utils";
+import { isSnapStore, isWindowsStore } from "jslib-electron/utils";
 
 import { IMenubarMenu } from "./menubar";
